refactored common timeline logic
