+++ src/Chirp.Web/Pages/PublicTimeline.cshtml
@@ -1,7 +1,7 @@
 @page "/"
 @using Chirp.Core
 @using Microsoft.AspNetCore.Mvc.TagHelpers
-@model Chirp.Web.Pages.PublicModel
+@model Chirp.Web.Pages.PublicTimelineModel
 @{
     ViewData["Title"] = "ITU-MiniTwit!";
     Layout = "Shared/_Layout";
